Text Mining project
"In this Text Mining project, we are examining the environmental policies of the five largest cities in France. Our goal is to measure similarities in environmental planning between 'smart' and 'non-smart' cities. The data and texts used in the project have been extracted from the websites of the relevant municipalities. The analysis of this data using text mining methods will help uncover potential similarities and differences in the environmental strategies of cities. This study aims to provide a valuable perspective on environmental sustainability."